requirements.txt
llama-index
openai
python-dotenv